ProjectExplorer: Guard against misuse of Node API
in QtC the only direct subclasses of Node are FileNode and FolderNode,
so that ensures that we always either have asFileNode() or
asFolderNode(), but custom subclasses could do that wrong. Warn instead
of crash in the FlatModel in that case (though there is no guarantee
that there are other places that make that assumption).

Change-Id: I29bc02116c775bc1b1068e791e08b2a7fa767917
Reviewed-by: hjk <hjk@qt.io>