Qml(Lite)Designer: Separate SingleAppInstances
Qt Design Studio has the option -qml-lite-designer that starts it in a
simplified mode, e.g. for opening individual files from Qt Creator. When
such a "lite" QDS is already running, we can and want to reuse it with
the "-client" option, but if a "normal" QDS is running we do not want
that to open instead.
So we need to separate the SingleApp IDs depending on the command line
option.

It is a bit ugly to hack the -qml-lite-designer into the global command
line handling, and it would probably be nicer to add a separate, generic
"-singleappid <id>" command line option that could be used by Qt Creator
when starting QDS. But that would require to wait for that new command
line option to roll out with QDS before we can use it in QtC. This can
still be added and switched to later when we assume that most QDS
installations have it.

Task-number: QTCREATORBUG-31005
Change-Id: Iad1d4ff7bf7ec7aa40cb2eae055ab8847a2e6730
Reviewed-by: hjk <hjk@qt.io>
(cherry picked from commit bfa0202814c872f1486fa20783823634cdfecd1c)
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>