QML Project Manager: fix host detection
Assuming that a kit that has deploy steps is not meant for
running on the host is not always correct.
When working with Qt for MCU for live previews there is an essentially
ordinary desktop kit with a QML exporter set to run as a deploy step.
We used to wrongly treat this as a remote setup, selecting
the standard QML runtime instead of QML puppet.
It caused caused fonts to be rendered incorrectly in Live Preview.

Amends: 8e0e097f559abd5322be674763b3baa29080c818

Fixes: QDS-14901
Change-Id: Iace1943bf888274fc47680a2bd8b1d787f665620
Reviewed-by: hjk <hjk@qt.io>
(cherry picked from commit ba2d4aaa52ff0e508db476537183e67c01f2a8ab)
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>