QmlProject: Allow opening individual .ui.qml files in "lite" QDS
QDS nowadays provides an option "-qml-lite-designer" that starts it in a
"lite" mode.

When opening QDS from Qt Creator, use the "normal" QDS when a
.qmlproject file is found. Otherwise start it with "-qml-lite-designer".
Switches to Process::startDetached with "-client" and "-qml-lite-
designer" on macOS too, since "open" wouldn't work for instance sharing
depending on the "-qml-lite-designer" option that we can do with the "-
client" option.

Task-number: QTCREATORBUG-31005
Change-Id: Ia9e00e0ed5ecad8c3c383ee46a6fd2594ee7b2e7
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>