QmlProject: Add a toolbutton for opening .ui.qml file in QDS
Adds a tool button to the text editor for .ui.qml files. A global action
is added too, that the user can assign a keyboard shortcut to.

Also adds a setting for configuring Qt Design Studio to the QML/JS
Editing settings, that is only shown if QmlProjectManager makes it
visible when QmlDesigner is not enabled. The setting defaults to the QDS
that is installed alongside Qt Creator, but the user can point it to an
installation, e.g. if QDS is installed through other sources, or QtC is
not from an online installer. Most importantly, the setting is a place
where this can be explained to the user.

This setting is also used for the Design mode landing page and the Open
with > Qt Design Studio functionality.

Fixes: QTCREATORBUG-31005
Change-Id: I86fbe205a4079e8d1192051eec3e38701d3c1dc0
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>