QbsBuild: Add support for libarchive
Went for the pragmatic way of making libarchive necessary
for building with qbs instead of trying to recreate the
CMakeLists and allow building the internally provided
sources.
On Linux and macOS libarchive and its development headers
are needed (system, brew,..).
On Windows an installation of libarchive is needed.

Set environment variable LIBARCHIVE_INSTALL_ROOT to the
base install folder if it cannot be found automatically.

Change-Id: Idbda1464fadb6b6c85f116f5f2a443a0f759b249
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>