Filter out keys from the attribution file that old Qt tools can't read
The qt_attributions.json file might contains keys that can't be parsed
by older qtattributionsscanner versions, but which are desired for SBOM
generation. Two such keys are: PURL and CPE.

To ensure documentation generation continues to work with older
versions of the tool when these keys are present, we now use
configure_file() / file(CONFIGURE) to copy the attribution file to
the build dir, but remove lines containing these keys.
We then pass the build dir attribution file path, rather than the
source dir one, to the tool.

We also replace any occurrences of @SOURCE_DIR@ with the absolute path
to the attribution file parent directory. This is needed so that the
build dir attribution file uses absolute path references to the
source dir licenses, so that the attribution scanner can refer to them
even if the attribution file is placed in the build dir.

Finally we annotate all the paths in the attribution file with the
@SOURCE_DIR@ token.

Change-Id: Ie99aa6e391ddc0b36122b485fda0499c3e795026
Reviewed-by: Eike Ziller <eike.ziller@qt.io>