qqmljs: adapt paths to .qrc files
The .qrc files have changed their locations in the build folder and
introduced a regression in the embedded code model. Make the algorithm
searching for .qrc files more relaxed such that it can find them again.

This fixes the embedded code model going to the build folder file
instead of the source folder file.

Fixes: QTCREATORBUG-32652
Change-Id: I42f86ecd7416c615629d45b3bf7bd2c89b1247dd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>