Utils: Fix a race condition in file path watcher
It was possible to lose the tracking of the file on saving.
Saving goes through the "safe save" mechanism, writing to a temporary
and then replacing the original with it.
When this happens file system watcher might detect that the file
was removed and stop watching it.
The attempt to re-add the file to the watcher again was prone to fail -
sometimes the check for the file existence was executed too early,
while the file was not yet fully replaced.
These changes ensure more consistent order of operations by
moving the file existence check out of the GlobalFileWatcher thread,
effectively postponing it.

Fixes: QDS-15036
Change-Id: Ic6fa489356fca3f9c6763af79d267c9b02c8d3ab
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>