QmlDesigner: Do not open extra editor when opening an example
Opening an example triggered explicit opening of the editor for the
main qml file. Code comment indicated that this was a workaround,
which appears to no longer be necessary, as project opens fine without
it. This workaround led to a duplicate DesignDocument instance being
created. This duplicate was not properly cleaned up at project close,
causing crash.

Fixes: QDS-15016
Change-Id: I7ea258470399edeffd22ed2d54f79f1779a0f8e7
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>