Fix debugging for Boot2Qt (isRemoteEngine)
Amends dbfa9e2d10a5c1e3e22b6134faead7a6f15e789e

That replaced the QUrl -> QString conversion from
    QString("%1:%2").arg(url.host()).arg(url.port());
to
    url.toString()
which is NOT the same.

Finally that leads to calling
    target remote tcp://<host>:<port>
intead of
    target remote tcp:<host>:<port>
which then fails.

The patch takes the stringified URL and in the isRemoteEngine
(StartRemoteProcess or AttachToRemoteServer) case changes the :// to
just :.

Fixes: QTCREATORBUG-32669
Change-Id: Ie551e3325f5854887a713ec75cb5b50078bb4ed6
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>