QmlDesigner: Do not remove imports is parsing virtual files
Since we use the same url as the QML file currently edited, if we remove
imports this breaks the imports.

This is because when parsing the "virtual file" imports could be removed
(which is disabled by this patch).

The verification step triggered by the merger does not
check for "temporarily" removed imports, because this was
done on a different model. The project storage cannot handle this.

The alternative would be to use a different "virtual" file URL.
This patch also sets all imports of the current document on the
template.

Task-number: QDS-15046
Change-Id: Id8826c28232eeba347af191f04a92872e1833c4e
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>