Lua: Only try loading plugins that are enabled
PluginManager::loadPluginsAtRuntime() will not load plugins
that are not "effectivelyEnabled" but "kill" them as they are
present in its "plugins" parameter, but are not loaded since
they are effectively disabled.

This led to all lua plugins that are not loaded on startup
to be "killed" so they could not be soft loaded later.

Change-Id: I929e0c44394d49dfa1df4646b1fa76e12e56f7c4
Reviewed-by: Christian Stenger <christian.stenger@qt.io>