Docker: Capture Environment from entry point
Docker containers often come with custom entry point scripts that inject
Environment variables into the shell. Since our bridge and other custom
processes are not spawned via the entry point they don't automatically
get access to the variables, breaking the KitDetector or the Terminal if
the user hopes to have them set.

As a workaround, we run "printenv" when a device is created and store
the output in its settings. We then use this environment as the device
environment when creating processes.

For old configs, we fetch the environment in fromMap in case it is not
set.

Fixes: QTCREATORBUG-32135
Change-Id: I16c04d391907698fd9f66fc3d33d068accb8054a
Reviewed-by: hjk <hjk@qt.io>