Layouting/Flow: Fix sizeHint and row size calculation
The right-most x coordinate of an item is "x + width - 1", so that must
be used for checking if the item is too large to fit into the row, not "x
+ width".

Change flow layout's size hint to fit all the items. If the flow layout is
allowed to request that width, then it should get it. That can still be
ignored by the using code, by setting an appropriate sizePolicy.

Change-Id: I3bc15d5683ffe2de7d835a757893784d5cef5b88
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>