Build: Simplify add_translation_targets
Deduplicate the _create_ts_custom_target calls by splitting the creation
of .ts and .qm targets. For both targets, one can specify a different
set of languages by passing TS_LANGUAGES or QM_LANGUAGES.

This is currently used for the "en" language that merely serves as a
typo hotfix translation.

Change-Id: I83bd10033188d4adf3dc0f78ecd81bcafc68ccaa
Reviewed-by: Eike Ziller <eike.ziller@qt.io>