Fix attaching QML Profiler to running application
The QML channel to use (the port that the user specified in the dialog
for attaching) was no longer set for the QmlProfilerRunner. That broke
with a22e79f38c13b652ad9639ee810647fc0c0e96b0
Avoid the usage of the ports gatherer with QmlProfilerRunner for the
attachToWaitingApplication case and set the qmlChannel manually.

Also avoid soft assert if there is no active run configuration.

Fixes: QTCREATORBUG-32617
Change-Id: I608cc736ed82284afbdd40277a54237f80f30306
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>