Git: Fix linkification of files with hex numbers
We don't want `foo-12345678.md` to be recognized as
as git hash `12345678`.

We cannot simply change `\b` to `\s` as the latter
would be included in the link. Therefore, switch to
lookbehind and lookahead.

Further, git shows output like `Amended "1245678"`,
which we do want to linkify and must be handled too.

Fixes: QTCREATORBUG-32740
Change-Id: I669473fe7b660975f182b3319565ff6672a2eb4c
Reviewed-by: Orgad Shaneh <orgads@gmail.com>