ActionManager: Work around crash on shutdown
The specific case is that RunControls can be deleted with deleteLater
(RunControlPrivate::checkAutoDeleteAndEmitStopped), which posts the
deleteLater on the event queue. If that happens "just before" shutdown,
the remaining events in the event queue are handled just before the
event loop exits, but _after_ aboutToQuit is sent and handled.
We ramp down plugins on aboutToQuit though, which means that we first
delete all plugins, which deletes ActionManager, and after that stray
deleteLater events are handled, ~RunControl > ~DebuggerRunTool >
~DebuggerEngine > ActionManager::unregisterAction.

Change-Id: I64f7901a647dc44cc88392312e9548cb46c4c192
Reviewed-by: hjk <hjk@qt.io>