CMakePM: Fix crash when adding files to target having zero arguments
Qt Creator restores a backup of a project's structure when a CMake
configuration fails.

If a user removes all arguments from a function defining a target,
configure CMake and have the backup restored can lead to a state where
the backup target exists in the project structure.

Then when trying to add files to this target Qt Creator crashes.

Fixes: QTCREATORBUG-32745
Change-Id: I9db9b54ae883f6c5b40da623a9984e9c33195d30
Reviewed-by: Eike Ziller <eike.ziller@qt.io>