QmlDesigner: Fix node order in QmlRewriter
This patch fixes the case for prepending an item to NodeListProperty

For the zero index corner case, If the order is available, we should
prepend the node. Otherwise, it should be appended.
This works for the case of NodeListProperty.

If the node order has not been changed, or is not important, we
should just append objects. This case works when we load and append
nodes consecutively, and node orders are not changed in the
transaction.

nodeLocation is optional. If it has a value, it means that we need
to consider the node location. Otherwise, we should only append.

When we search for the items to insert an object after them, if we
want to insert an object to the top (pos = -1), we should find the
current first object, and return the last non-object item found
before the first object.
If there's no object, it will be automatically return the last
non-object.

Fixes: QDS-15226
Change-Id: I5c5c23a27a956baa15a7a4eecae2fa5d4f9bb650
Reviewed-by: Knud Dollereder <knud.dollereder@qt.io>