QmlDesigner: Fix DesignSystemView for project storage
The projectStorageDependencies are only available if actually a project
is opened. DSStore is part of a library and also used in the importer.
Instead of changing DSStore, we create the DSStore in the first attach
after a project was opened.

DesignSystemInterface is the interface to the QML UI. We guard access
in case no DSStore is created, yet.

TODO: We should only load the DesignSystem on demand if the view is
actually active.

Task-number: QDS-15070
Change-Id: I6f52375d029e575582dc8f16abc50c157666f287
Reviewed-by: Marco Bubke <marco.bubke@qt.io>