Add an UpdateInfo::Service and use it for a Install QDS button
Adds an Update/Installer API that other plugins can use to trigger
installation of a Qt Online Installer component.

The API takes a component name/regex. It opens the dialog and uses the
headless Qt IFW commands to

- finds components matching the pattern
- shows the user details on which components are available, or if none
  are available
- asks the user if they should be installed
- installs the components (showing the installer output)

The "Install QDS button" is available if Qt Design Studio is not found,
and uses the new API to allow the user to install it, without the need
to manually open MaintenanceTool, finding the right component, etc.

Fixes: QTCREATORBUG-30787
Change-Id: I590bb557c93373bcc73ab70524720b00c05763e5
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>