Debugger: Wait for terminal to finish after finishing debugger
When running process uses terminal, and the process is interrupted,
the d'tor of such process hangs on TerminalInterface::killStubProcess().
This shouldn't happen, and it should be addressed separately.

As a workaround we don't destroy the running terminal process,
but stop debugger first and wait for terminal to finish.

Amends 6d9575f942289fbf5312f74f560e327e9dc5a378

Fixes: QTCREATORBUG-32721
Change-Id: I9612924973cfe97848758c3c688b9dbf20534a4d
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>