QML Designer: Load relevant fonts in MCU projects
For MCU projects, QDS will use fonts that are closer to those that are
used in QUL. This mostly applies to the cases of implicitly assumed
fonts (i.e. there's no user-specified font properties).
In this case text items will use the font set in
`MCU.Config.defaultFontFamily` property, and, if this property is not
set, DejaVu Sans, which corresponds to the behavior of Qt for MCUs.
Also only the relevant fonts will be suggested in the font picker
combobox. These come from the MCU installation and replace the standard
set of fonts (Arial, Courier, Times New Roman, etc.) in MCU projects.

From a technical perspective:
  * added a new model - `FontResourceModel` - as font handling logic is
    now too complex for a more generic `FileResourceModel`.
  * naturally, some wiring to be able to access MCU font throughout
    various parts of the project.

Fixes: QDS-10597
Fixes: QDS-15079
Fixes: QDS-15156
Change-Id: I25aa4b0f89cca864f8fd41e0812f448fe307b104
Reviewed-by: Knud Dollereder <knud.dollereder@qt.io>