Core: Replace IDocument::OpenResult by Utils::Result<>
After the preceding change, the semantics were the same.

The need to specify new error messages in some places indicates that
errors were not very descriptive in some places before. For now,
use some untranslated strings in these cases, the idea is to centrally
bundle them via constants to ResultError similar to
ResultNotImplemented.

Change-Id: I6388cc43e0a540068ea73117a91fa2db982fbefa
Reviewed-by: Christian Stenger <christian.stenger@qt.io>