McuSupport: fix optional packages
MCU plugin didn't have to handle "optional" property in MCU kit descriptions.
I believe this is mainly because "optional" was always "false" for all
the previous MCU targets. A newly introduced kit, with one package set
"optional" to "true", manifested the issue. The user couldn't create a
kit without providing a valid path to that optional package.

The user should be able to leave the optional package path empty or
provide a valid path.

This change introduces a new Groupbox widget to display the optional
packages. An optional package with empty path is considered valid.

Task-number: QTCREATORBUG-32777
Change-Id: I5f27952efc77bc5e4351ab4c013669eb6056810f
Reviewed-by: Sivert Krøvel <sivert.krovel@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>