Utils: Fix BaseAspect enabler usage
Previously the enabler was only used indirectly. This meant that
cancelling (e.g. in the settings dialog) would not be registered by
the enabled aspect. On the next open of the settings dialog the
m_enabled would still be on the previous state, leading to a UI
that is out of sync.

Change-Id: I19c60063e58b831504bf0aedfdd71d983e55ffb9
Reviewed-by: hjk <hjk@qt.io>