Coco: fix crash in CocoBuildStep
The bug was reported via Sentry and I could not reproduce it here, but
the immediate cause was clear: m_buildSettings was used when it was
still a null pointer - certainly because CocoBuildStep::display() was
not yet called.

I have added code that prevents this.

Fixes: QTCREATORBUG-32850
Change-Id: I8226e1466870e77749aff276de150b79f75332a4
Reviewed-by: David Schulz <david.schulz@qt.io>
(cherry picked from commit c052dda76b3b3e64f13a39cc1adaec91b6195944)