QmlDesigner: Fix crash when switching to first .ui.qml file
if Qt Creator is already in Design mode.

QmlDesignerPlugin::showDesigner() was only called on a switch to Design
mode while the editor is a .ui.qml file. So it wasn't called when opening
a .ui file in Design mode and then switching to a .ui.qml file. But
showDesigner() creates the QmlDesigner widgets, so it must be called.

Fixes: QTCREATORBUG-32854
Change-Id: I02a498ce00fc1afdee900752ade7af6c62b787b7
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>