Git: Return Result<QString> from GitClient::synchronizedLog()
Plan is to simplify the overall logic with less "shared"
out-parameters.

Complication in this change here is the magic in msgCannotRun() to
show or not show the error immediately, depending on whether
errorMessage is nullptr or not.

Since we now always pass a non-null errorMessage argument, some
VcsOutputWindow::appendError(...) calls have to be moved to the
caller side.

Change-Id: Ibd20d7ba0811ec34f86dc96914d4fbe7ae446ada
Reviewed-by: David Schulz <david.schulz@qt.io>