TextEditor: improve tab settings auto detection
Instead of just collecting the indentation for the checked blocks we now
collect the indentation delta between blocks. If a sufficient amount of
common deltas have been found and the most common indentation is a
mutiple of the delta we base the tab settings on the delta instead of
the smallest indentation.

Fixes: QTCREATORBUG-32861
Change-Id: I7d7c5eac6b647671a351c6eaa7527af12792e257
Reviewed-by: Christian Stenger <christian.stenger@qt.io>