QmlDesigner: Fix drag & drop 3D components into 2D/3D view
The created components do not have metadata files that contain
the node's hints. With their help, a decision is made whether
the item can be placed in the view. The fix is to add two hints that
allow to create an item in the proper view.

In addition, a metainfo has been added to check if an object
is based on Object3D. This allows to check if the element is from QtQuick3D.

This fix solves issues for non-project storage builds only.

Task-number: QDS-15237
Pick-to: qds/4.7
Change-Id: Ib00c26989b0400d3aad48882f5c9f8fc1b5bef40
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>