QtSupport: Add another hack to back-map paths of .qml files
Paths of .qml files go through several location before ending up in
the binary (source directory, build directory, entries in .qrc files).

The qml debug server side does not do any back-mapping to source
directory files when reporting e.g.  stack frames or breakpoint
locations, so this a FileInProjectFinder is supposed to do the work,
some deterministic, some by guessing.

The most deterministic way is to read .qrc fils, but in modern Qt 6.x(?)
CMake projects, .qrc files containing .qml file are typically on-the-fly
and are not part of the project's sources anymore.

On top of this, the heuristic to search for the files in the project
directory which would normally help as fallback fails for in-source
builds due to some unfortunate naming of build artifacts (for a fie
called "testprojects/untitled/Main.qml" in the source, a file
"testproject/untitled/untitled/Main.qml" build artifact will be generated
/and/ found by the back-mapping heuristics using "prefixToIgnore" in
FileInProjectFinder::checkProjectDirectory.

To work around further, we add the relevant generated .qrc files manually
in that case.

Change-Id: Idcab1a3a5a86bd055a8b98e7fec13eb59d403b1e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>