InfoBar: Unify code for hiding / suppressing info on button click
Many buttons that are added to info bar entries should also hide the
info bar when clicked, and many should even suppress the info bar from
showing again (persistently or for the session).

The code for that is duplicate in all the button callbacks, which is
error prone and also globally retrieves the info bar again on button
click, which makes assumptions that we want to break in the future
(ICore::infoBar() can return a different info bar, depending on
settings, which can change during runtime).

Add an additional enum parameter to InfoBarEntry::addCustomButton() that
unifies that behavior at a central place, and also unify the button click
handling for the InfoBarDisplay implementations.

This actually fixes a bug with hiding the "Update" info bar which used an
outdated and wrong ID for the duplicated hiding code.

Change-Id: I20cd5b5b566e817a873daa6fb77db25eefcb32f0
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>