RFC: qmloutline: disable drag-and-drop
The qml outline has been broken for quite some time now. Remove the
drag-and-drop functionality.

Rationale:
Drag-and-dropping code using the outline is pretty unusual. For
example, drag-and-drop is not possible via the C++ code outline in Qt
Creator, or other editors like VS Code, and even the LSP protocol
offers no way to provide such a functionality.

In addition, drag-and-drop has been broken since QtC 6 and I don't
think that people actually still (try to) use this feature.

Also, we might want to replace the embedded qml codemodel of QtC with
qmlls in the future, and implementing a custom LSP extension to support
drag-and-drop in the outline for feature-parity with a broken and unused
feature currently seems like a lot of effort for a broken and unused
feature.

For future reference: the drag-and-drop functionality stopped working
because the QDataStream uses a "long long" to write an index into
Utils::DropMimeData, but uses an "int" to read the index back.
Reordering code via the "fixed" outline is a really weird experience:
you can reorder functions but only if there are no property definitions
in the Qml code, properties can only be moved all the way to the top or
all the way to the bottom, sometimes no re-ordering happens, and
sometimes the function or property gets moved at the wrong place in the
qml code.

Fixes: QTCREATORBUG-26391
Change-Id: I14cfe59532f0175b821e589327156d154a632d2b
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>