macOS: Fix configuration issues with CMake 4 and later
CMake 4 significantly changed its behavior with regards to setting up the
CMAKE_OSX_SYSROOT - it doesn't set it anymore, assuming that the
compiler will be from the command line tools /usr/bin/clang(++) which
set the sysroot to the latest macOS themselves. For details see
https://github.com/Kitware/CMake/commit/7b195312919cbf69

So it turns out that the choice for unification of the chosen compiler to
the Xcode.app ones (as detected from the iOS plugin) in
f49f9b43da38be4ec0f46fe6cd6930467c7444c1 bites back now.

On the contrary, deprioritize the Xcode.app compilers, so the /usr/bin
ones are used for Desktop kits. The auto-generated iOS kits keep
choosing the Xcode.app compilers and set their sysroot explicitly
already.

Fixes: QTCREATORBUG-32877
Change-Id: Ia8a4aab009d4d959743a27091217ea9b80811da0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>