LSP: mark some clients as not activatable
Differentiate between normal lsp clients that can be used for tasks like
completions/outline/highlight and those that just uses the protocol to
have a basic document update communication and uses their own messages
for specialised tasks like the copilot, coco or ai assistant server. The
user should not be able to manually switch to any of those servers as
the active server for the current document, so filter those servers in
the editor tool bar.

Task-number: QTCREATORBUG-32909
Change-Id: Idd5e1c10e613ce986a56defad02a3787aeae85b2
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>