Core: Hack in ManhattanStyle for distinguishable button highlight
The background of highlighted view items and the outline of highlighted
buttons are (at least in Qt's Fusion style) both painted with the
QPalette::Highlight role. The Figma design for the "2024 themes" defines
the background of the highlighted view item to be gray, which results in
a gray and therefore indistiguishable highlighted button.

This hack uses ManhattanStyle::polish to check whether
QPalette::Highlight and QPalette::Accent roles differ, which is an
indicator for "2024 themes" and assigns QPalette::Accent as
QPalette::Highlight for QAbstractButtons.

Fixes: QTCREATORBUG-32549
Change-Id: Ib0cae4948c565d547f2d04a709c73450299ee563
Reviewed-by: Cristian Adam <cristian.adam@qt.io>