ExtensionManager: Fix browser item rendering for Qt 6.9.0
A QItemDelegate in the ExtensionManager widget holds a QWidget and
renders its content using QWidget::render(). Before the rendering, it
sets various texts and pixmaps of the the widget's child widgets. The
child widgets are laid out by Q*Layout.

This method used to work fine until the Qt Creator package builds
switched to Qt 6.9.0. With Qt 6.9.0, the sizes of the child widgets did
not update anymore after setting label texts and calling
QWidget::render().

This change calls QLayout::invalidate() of a specific child widget
before rendering, which works around the issue for Qt 6.9.0.

Fixes: QTCREATORBUG-32954
Change-Id: I038fdf1407a66134bf8bb6427c5a42ad6d367b5f
Reviewed-by: Eike Ziller <eike.ziller@qt.io>