Utils: Speed up ChangeSet::apply()
The previous algorithm traversed the remaining list of operations on
every application of an operation, incurring quadratic effort.
This was done because the positions of the other operations potentially
need updating if they are located after the current one in the file.
We solve this by sorting the list by position and applying the
operations starting at the back, with some extra logic for operations
with the same position. The cost is now linear for real-world cases.
Example benchmark: Applying ~5000 operations (from clang-tidy) to a file
used to take four seconds and now takes one second.

Task-number: QTCREATORBUG-25394
Change-Id: Ie386ddbc12696b67406b37e67f52072d9c1bf6a0
Reviewed-by: David Schulz <david.schulz@qt.io>