qmljs formatter: take tab setting into account
Prior to this commit, built-in qml formatter was always inserting spaces
even if tabSettings is in use-tabs mode only when "Reformat Document" action
invoked. Auto format on save was working correct because it rewrites the
document on save() taking the tab settings into account.

To fix this, indent the blocks after formatting is performed.

Also, apply the current codestyle before formatting performed. This
makes sure the current tab settings are applied to document.

Task-number: QTCREATORBUG-32951
Task-number: QTCREATORBUG-32963
Change-Id: Ia862f607cafa40078a25beb4b06e67b7c7ca8b94
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>