Docker: Fix done emissions in case of failed process start
In case that the first line received from the process is not starting with
"__qtc" (the remote process id marker) we need to fail the startup of the process.

In this case the DockerProcessImpl would emit "done" twice, once from the
startup check, and once from the process actually exiting.

Instead, we remember the first line, and just force the process to quit, relying
on the final "done" signal to actually emit the done signal.

Fixes: QTCREATORBUG-32960
Change-Id: I81f54571b1eb83676bf30571e630dd3ef1ef9f8c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>