Welcome: Remove the "Qt Creator" label
The initial "Welcome to Qt Creator" that was meant to send a friendly
welcoming message to the user got shortened to "Qt Creator" which
does not serve a reason anymore as people usually know which application
they are using, or can it look up in the window title in most systems.

The proposal to re-add the "Welcome to" did not find consensus, so
remove the label instead.

Amends 7b3886d22e.

Change-Id: Icd918d813e5aa41e5cf14813a1809c3a8b929d2f
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>