Utils: Attach lifetime of guardedCallback to Guard object
This changes Utils::guardedCallback() such that the passed in callback
object will be deleted once the guard object dies. This is necessary to
bind the lifetime of captured variables in a lambda to the guard object.

Otherwise the captures might be deleted after the guarded object has
long been destroyed which is a problem especially for Lua variables that
need the Lua context to be alive during their destruction.

This allows us to remove the hack/workaround in InfoBarDisplay::update
that used QObject::disconnect(...) to get rid of the connected lambdas
immediately.

Change-Id: Icd4e187d0fdfed6a6fb9a0d890c43411c104ee55
Reviewed-by: Eike Ziller <eike.ziller@qt.io>