Fix the selection in Open Documents view
There were two cases where the view behaved weird: When clicking on the
close icon with the mouse, and when giving the view focus (e.g. with the
shortcut for the view) and navigating by keyboard and closing documents
with the delete key.

The main issue is that the tree view selects the item that is clicked
on, even if the click is in the second column, which is the column with
the X close button. If that happens, the item that is selected _after_
closing and removing the item from the tree, is the item below the
removed item, which is not what we want. If the user clicks on the X
button, we do not want the selection to change at all (if the closed
item itself isn't the selected item).

The code had a brute force attempt for "fixing" that by setting the
selection to the current document in the editor after an item was
closed. Which both leads to scrolling to that item, which we do not
want, and if the user closed with the delete key while the view had
focus, that also completely moves the focus item there.

Instead, prevent the tree view base implementation from selecting the
clicked item, if that happened in the second column, by handling that
case completely ourselves. Also sync the view to the current document
when focus changes away from the Open Documents view, to get syncing
back after a user switched focus to the view and possibly moved the
selection around.

Fixes: QTCREATORBUG-32586
Change-Id: Ibb170a3f3ce2e07ebf607dd029cf26a45fe851ce
Reviewed-by: Christian Stenger <christian.stenger@qt.io>