QmlDesigner: Keep global transform on reparenting in navigator
By default, if user reparents a 3D node from under one 3D node to
another, the transform of the reparented node is now adjusted so that
after reparenting it will have the same global position, orientation,
and size as before.

This is only possible to do reliably if the old and new ancestor nodes
do not have both non-uniform scaling and rotation, as combination of
those can skew the transform into such as state that cannot be
replicated with simple three axis scaling. In case both non-uniform
scaling and rotation are present in the ancestral chains, the local
scaling will not be adjusted - just the position and rotation.

If reparented node is part of timeline animation in either new or old
location, transform is not adjusted.

If user holds down Alt key during drop, transform is not adjusted.

Since reparenting is done in base state, transform adjustment is
also always done in the base state.

Fixes: QDS-15392
Change-Id: Ibfda7459008aa0215d5cb368114f84a494e3e4b3
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>