QmlJS: Replace FileSystemWatcher with FilePathWatcher
FileSystemWatcher would create a map with the files and at insertion in
the map it would do a file access.

Move to have a map of FilePaths to FilePathWatchers, which would not do
file access at map creation.

This way of watching files is similar to what Core::DocumentManager is
doing.

This works way better. There is no freeze when a project using 500 qml
plugins is loaded, and also when the project is rebuilt.

Fixes: QTCREATORBUG-32209
Change-Id: Ia1be55e316f163a8c1059f40b855fbd6cbc06fdd
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>