Utils: Add SpanAll helper
The Span modifier originally only added the span attribute
to the last item added from its "items" parameter. In some ways
that is wrong, as you would expect it to be added to all the items
you provide. This behavior is however (mis)used when adding
for instance an aspect to a layout which will add a label and
an input widget at the same time. Then it is beneficial if the span
is only applied to the input element, and not to the label.

As a workaround, this patch restores the old behavior and adds a
new modifier "SpanAll" that applies the span to all of its items.

Amends 515d5dffa4b2d461a74b67aa41c99227005be188

Change-Id: I20f8ca4deeab0d8de283725b3b7af19396a710f7
Reviewed-by: hjk <hjk@qt.io>