ExtensionSystem: Fix that IsoIconBrowser indirectly enabled QmlDesigner
Normally, when a plugin is enabled, all its dependencies are indirectly
force-enabled too. This logic was added when we ran into the situation
that in a new version of Qt Creator an existing plugin A got a new
dependency on another existing plugin B, and if users previously
disabled plugin B, plugin A would silently not load, leading to broken
features without any hint on what was going on.

Now we do have a new situation: QmlDesigner's functionality is extended
by plugins, but QmlDesigner is disabled by default in Qt Creator. So
far, all the plugins that extend QmlDesigner were either also disabled
by default (and the whole set was explicitly enabled for the Qt Design
Studio installation), or they were not Qt Creator plugins, but managed
by a separate "QmlDesigner plugin manager". The latter has changed, and
the IsoIconBrowser plugin that extends QmlDesigner was changed to a
"regular" Qt Creator plugin.

This leads to the effect that if IsoIconBrowser is enabled by default,
it forces QmlDesigner to be enabled, which we do not want. But if
IsoIconBrowser is disabled by default, users who enable QmlDesigner also
need to enable IsoIconBrowser manually to get that functionality (and
users probably don't know about that).

Disable the "indirectly force-enable dependencies" mechanism for
dependencies that are disabled by default.

Fixes: QTCREATORBUG-32907
Change-Id: I97e279d9ac8a252fed260b3ffc7dd7ebc0135d7d
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>